Issue: dosen't display comments
