+++ portfolio/src/main/java/com/google/sps/servlets/BlobstoreUploadUrlServlet.java
@@ -33,7 +33,7 @@ public class BlobstoreUploadUrlServlet extends HttpServlet {
   @Override
   public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
     BlobstoreService blobstoreService = BlobstoreServiceFactory.getBlobstoreService();
-    String uploadUrl = blobstoreService.createUploadUrl("/my-form-handler");
+    String uploadUrl = blobstoreService.createUploadUrl("/comment");
 
     response.setContentType("text/html");
     response.getWriter().println(uploadUrl);
