+++ portfolio/src/main/java/com/google/sps/servlets/FormHandlerServlet.java
@@ -1,101 +1,101 @@
-// Copyright 2019 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// // Copyright 2019 Google LLC
+// //
+// // Licensed under the Apache License, Version 2.0 (the "License");
+// // you may not use this file except in compliance with the License.
+// // You may obtain a copy of the License at
+// //
+// //     https://www.apache.org/licenses/LICENSE-2.0
+// //
+// // Unless required by applicable law or agreed to in writing, software
+// // distributed under the License is distributed on an "AS IS" BASIS,
+// // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// // See the License for the specific language governing permissions and
+// // limitations under the License.
 
-package com.google.sps.servlets;
+// package com.google.sps.servlets;
 
-import com.google.appengine.api.blobstore.BlobInfo;
-import com.google.appengine.api.blobstore.BlobInfoFactory;
-import com.google.appengine.api.blobstore.BlobKey;
-import com.google.appengine.api.blobstore.BlobstoreService;
-import com.google.appengine.api.blobstore.BlobstoreServiceFactory;
-import com.google.appengine.api.images.ImagesService;
-import com.google.appengine.api.images.ImagesServiceFactory;
-import com.google.appengine.api.images.ServingUrlOptions;
-import java.io.IOException;
-import java.io.PrintWriter;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.util.List;
-import java.util.Map;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+// import com.google.appengine.api.blobstore.BlobInfo;
+// import com.google.appengine.api.blobstore.BlobInfoFactory;
+// import com.google.appengine.api.blobstore.BlobKey;
+// import com.google.appengine.api.blobstore.BlobstoreService;
+// import com.google.appengine.api.blobstore.BlobstoreServiceFactory;
+// import com.google.appengine.api.images.ImagesService;
+// import com.google.appengine.api.images.ImagesServiceFactory;
+// import com.google.appengine.api.images.ServingUrlOptions;
+// import java.io.IOException;
+// import java.io.PrintWriter;
+// import java.net.MalformedURLException;
+// import java.net.URL;
+// import java.util.List;
+// import java.util.Map;
+// import javax.servlet.annotation.WebServlet;
+// import javax.servlet.http.HttpServlet;
+// import javax.servlet.http.HttpServletRequest;
+// import javax.servlet.http.HttpServletResponse;
 
-/**
- * When the user submits the form, Blobstore processes the file upload and then forwards the request
- * to this servlet. This servlet can then process the request using the file URL we get from
- * Blobstore.
- */
-@WebServlet("/my-form-handler")
-public class FormHandlerServlet extends HttpServlet {
+// /**
+//  * When the user submits the form, Blobstore processes the file upload and then forwards the request
+//  * to this servlet. This servlet can then process the request using the file URL we get from
+//  * Blobstore.
+//  */
+// @WebServlet("/my-form-handler")
+// public class FormHandlerServlet extends HttpServlet {
 
-  @Override
-  public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
+//   @Override
+//   public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
 
-    // Get the message entered by the user.
-    String message = request.getParameter("message");
+//     // Get the message entered by the user.
+//     String message = request.getParameter("message");
 
-    // Get the URL of the image that the user uploaded to Blobstore.
-    String imageUrl = getUploadedFileUrl(request, "image");
+//     // Get the URL of the image that the user uploaded to Blobstore.
+//     String imageUrl = getUploadedFileUrl(request, "image");
 
-    // Output some HTML that shows the data the user entered.
-    // A real codebase would probably store these in Datastore.
-    PrintWriter out = response.getWriter();
-    out.println("<p>Here's the image you uploaded:</p>");
-    out.println("<a href=\"" + imageUrl + "\">");
-    out.println("<img src=\"" + imageUrl + "\" />");
-    out.println("</a>");
-    out.println("<p>Here's the text you entered:</p>");
-    out.println(message);
-  }
+//     // Output some HTML that shows the data the user entered.
+//     // A real codebase would probably store these in Datastore.
+//     PrintWriter out = response.getWriter();
+//     out.println("<p>Here's the image you uploaded:</p>");
+//     out.println("<a href=\"" + imageUrl + "\">");
+//     out.println("<img src=\"" + imageUrl + "\" />");
+//     out.println("</a>");
+//     out.println("<p>Here's the text you entered:</p>");
+//     out.println(message);
+//   }
 
-  /** Returns a URL that points to the uploaded file, or null if the user didn't upload a file. */
-  private String getUploadedFileUrl(HttpServletRequest request, String formInputElementName) {
-    BlobstoreService blobstoreService = BlobstoreServiceFactory.getBlobstoreService();
-    Map<String, List<BlobKey>> blobs = blobstoreService.getUploads(request);
-    List<BlobKey> blobKeys = blobs.get("image");
+//   /** Returns a URL that points to the uploaded file, or null if the user didn't upload a file. */
+//   private String getUploadedFileUrl(HttpServletRequest request, String formInputElementName) {
+//     BlobstoreService blobstoreService = BlobstoreServiceFactory.getBlobstoreService();
+//     Map<String, List<BlobKey>> blobs = blobstoreService.getUploads(request);
+//     List<BlobKey> blobKeys = blobs.get("image");
 
-    // User submitted form without selecting a file, so we can't get a URL. (dev server)
-    if (blobKeys == null || blobKeys.isEmpty()) {
-      return null;
-    }
+//     // User submitted form without selecting a file, so we can't get a URL. (dev server)
+//     if (blobKeys == null || blobKeys.isEmpty()) {
+//       return null;
+//     }
 
-    // Our form only contains a single file input, so get the first index.
-    BlobKey blobKey = blobKeys.get(0);
+//     // Our form only contains a single file input, so get the first index.
+//     BlobKey blobKey = blobKeys.get(0);
 
-    // User submitted form without selecting a file, so we can't get a URL. (live server)
-    BlobInfo blobInfo = new BlobInfoFactory().loadBlobInfo(blobKey);
-    if (blobInfo.getSize() == 0) {
-      blobstoreService.delete(blobKey);
-      return null;
-    }
+//     // User submitted form without selecting a file, so we can't get a URL. (live server)
+//     BlobInfo blobInfo = new BlobInfoFactory().loadBlobInfo(blobKey);
+//     if (blobInfo.getSize() == 0) {
+//       blobstoreService.delete(blobKey);
+//       return null;
+//     }
 
-    // We could check the validity of the file here, e.g. to make sure it's an image file
-    // https://stackoverflow.com/q/10779564/873165
+//     // We could check the validity of the file here, e.g. to make sure it's an image file
+//     // https://stackoverflow.com/q/10779564/873165
 
-    // Use ImagesService to get a URL that points to the uploaded file.
-    ImagesService imagesService = ImagesServiceFactory.getImagesService();
-    ServingUrlOptions options = ServingUrlOptions.Builder.withBlobKey(blobKey);
+//     // Use ImagesService to get a URL that points to the uploaded file.
+//     ImagesService imagesService = ImagesServiceFactory.getImagesService();
+//     ServingUrlOptions options = ServingUrlOptions.Builder.withBlobKey(blobKey);
 
-    // To support running in Google Cloud Shell with AppEngine's devserver, we must use the relative
-    // path to the image, rather than the path returned by imagesService which contains a host.
-    try {
-      URL url = new URL(imagesService.getServingUrl(options));
-      return url.getPath();
-    } catch (MalformedURLException e) {
-      return imagesService.getServingUrl(options);
-    }
-  }
-}
+//     // To support running in Google Cloud Shell with AppEngine's devserver, we must use the relative
+//     // path to the image, rather than the path returned by imagesService which contains a host.
+//     try {
+//       URL url = new URL(imagesService.getServingUrl(options));
+//       return url.getPath();
+//     } catch (MalformedURLException e) {
+//       return imagesService.getServingUrl(options);
+//     }
+//   }
+// }
